Move circular_buffer and buffered_tx under hoverboard module.
