+++ cross/hoverkite-firmware/src/hoverboard/mod.rs
@@ -1,9 +1,10 @@
 mod buzzer;
 mod motor;
+pub mod util;
 
 pub use self::buzzer::Buzzer;
 use self::motor::{HallSensors, Motor};
-use crate::buffered_tx::{BufferState, BufferedSerialWriter, Listenable};
+use self::util::buffered_tx::{BufferState, BufferedSerialWriter, Listenable};
 use core::{cell::RefCell, mem, ops::Deref};
 use cortex_m::{
     interrupt::{free, Mutex},
