+++ cross/hoverkite-firmware/src/hoverboard/util/circular_buffer.rs
